+++ src/app/page.tsx
@@ -4,16 +4,17 @@ import { useEffect, useRef, useState } from 'react';
 import { CustomOverlayMap, Map } from 'react-kakao-maps-sdk';
 
 import BottomSheet from '@components/main/BottomSheet';
-import useLocation from '@hooks/useLocation';
+import useCoordinate from '@hooks/useCoordinate';
 import CurrentLocationIcon from 'public/assets/icon24/current_location_24.svg';
 
 export default function Home() {
   const mapRef = useRef<kakao.maps.Map>(null);
-  const { center, setCurrentUserLocation } = useLocation();
+  const { center, currentUserCoordinate, getCurrentUserCoordinate } =
+    useCoordinate();
   const [isBottomSheetShowing, setIsBottomSheetShowing] = useState(false);
 
   useEffect(() => {
-    setCurrentUserLocation();
+    getCurrentUserCoordinate();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
@@ -25,7 +26,7 @@ export default function Home() {
         className="w-full h-full"
         isPanto={true}
       >
-        <CustomOverlayMap position={center}>
+        <CustomOverlayMap position={currentUserCoordinate}>
           <CurrentLocationIcon />
         </CustomOverlayMap>
       </Map>
